requirements/constraints.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    make pip-compile
#
annotated-types==0.7.0
    # via
    #   -r requirements/main.txt
    #   pydantic
argcomplete==3.5.1
    # via
    #   -r requirements/dev.txt
    #   nox
asgiref==3.8.1
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   django
    #   django-stubs
attrs==24.2.0
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   interrogate
    #   jsonschema
    #   referencing
black==24.10.0
    # via -r requirements/dev.txt
build==1.2.2.post1
    # via
    #   -r requirements/dev.txt
    #   pip-tools
certifi==2024.8.30
    # via
    #   -r requirements/dev.txt
    #   requests
charset-normalizer==3.4.0
    # via
    #   -r requirements/dev.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/dev.txt
    #   black
    #   interrogate
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements/dev.txt
    #   interrogate
colorlog==6.8.2
    # via
    #   -r requirements/dev.txt
    #   nox
coverage==7.6.3
    # via -r requirements/dev.txt
distlib==0.3.9
    # via
    #   -r requirements/dev.txt
    #   virtualenv
dj-database-url==2.2.0
    # via -r requirements/main.txt
django==5.1.2
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   dj-database-url
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   drf-spectacular
django-stubs==5.1.0
    # via
    #   -r requirements/dev.txt
    #   djangorestframework-stubs
django-stubs-ext==5.1.0
    # via
    #   -r requirements/dev.txt
    #   django-stubs
djangorestframework==3.15.2
    # via
    #   -r requirements/main.txt
    #   drf-spectacular
djangorestframework-stubs==3.15.1
    # via -r requirements/dev.txt
drf-spectacular==0.27.2
    # via -r requirements/main.txt
filelock==3.16.1
    # via
    #   -r requirements/dev.txt
    #   virtualenv
gunicorn==23.0.0
    # via -r requirements/main.txt
idna==3.10
    # via
    #   -r requirements/dev.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/main.txt
    #   drf-spectacular
iniconfig==2.0.0
    # via
    #   -r requirements/dev.txt
    #   pytest
interrogate==1.7.0
    # via -r requirements/dev.txt
isort==5.13.2
    # via -r requirements/dev.txt
jsonschema==4.23.0
    # via
    #   -r requirements/main.txt
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/main.txt
    #   jsonschema
markdown==3.7
    # via -r requirements/main.txt
mypy==1.12.0
    # via -r requirements/dev.txt
mypy-extensions==1.0.0
    # via
    #   -r requirements/dev.txt
    #   black
    #   mypy
nox==2024.10.9
    # via -r requirements/dev.txt
packaging==24.1
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   black
    #   build
    #   gunicorn
    #   nox
    #   pytest
pathspec==0.12.1
    # via
    #   -r requirements/dev.txt
    #   black
pip-tools==7.4.1
    # via -r requirements/dev.txt
platformdirs==4.3.6
    # via
    #   -r requirements/dev.txt
    #   black
    #   virtualenv
pluggy==1.5.0
    # via
    #   -r requirements/dev.txt
    #   pytest
psycopg==3.2.3
    # via -r requirements/main.txt
psycopg-binary==3.2.3
    # via -r requirements/main.txt
py==1.11.0
    # via
    #   -r requirements/dev.txt
    #   interrogate
pydantic==2.9.2
    # via -r requirements/main.txt
pydantic-core==2.23.4
    # via
    #   -r requirements/main.txt
    #   pydantic
pyproject-hooks==1.2.0
    # via
    #   -r requirements/dev.txt
    #   build
    #   pip-tools
pytest==8.3.3
    # via
    #   -r requirements/dev.txt
    #   pytest-django
pytest-django==4.9.0
    # via -r requirements/dev.txt
python-decouple==3.8
    # via -r requirements/main.txt
pyyaml==6.0.2
    # via
    #   -r requirements/main.txt
    #   drf-spectacular
referencing==0.35.1
    # via
    #   -r requirements/main.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements/dev.txt
    #   djangorestframework-stubs
rpds-py==0.20.0
    # via
    #   -r requirements/main.txt
    #   jsonschema
    #   referencing
ruff==0.6.9
    # via -r requirements/dev.txt
sqlparse==0.5.1
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   django
tabulate==0.9.0
    # via
    #   -r requirements/dev.txt
    #   interrogate
types-dj-database-url==1.3.0.4
    # via -r requirements/dev.txt
types-pyyaml==6.0.12.20240917
    # via
    #   -r requirements/dev.txt
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.0.20241016
    # via
    #   -r requirements/dev.txt
    #   djangorestframework-stubs
typing-extensions==4.12.2
    # via
    #   -r requirements/dev.txt
    #   -r requirements/main.txt
    #   dj-database-url
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
    #   psycopg
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via
    #   -r requirements/main.txt
    #   drf-spectacular
urllib3==2.2.3
    # via
    #   -r requirements/dev.txt
    #   requests
    #   types-requests
virtualenv==20.26.6
    # via
    #   -r requirements/dev.txt
    #   nox
wheel==0.44.0
    # via
    #   -r requirements/dev.txt
    #   pip-tools
whitenoise==6.7.0
    # via -r requirements/main.txt

# The following packages are considered to be unsafe in a requirements file:
pip==24.2
    # via
    #   -r requirements/dev.txt
    #   pip-tools
setuptools==75.1.0
    # via
    #   -r requirements/dev.txt
    #   pip-tools
